cancer-screening-system-web/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python
pillow
pydantic
python-multipart
matplotlib